Moved FileUpload and WriteTemplate to the components directory
